frontend_streamlit/requirements_streamlit.txt
streamlit>=1.30.0
plotly>=5.18.0
httpx>=0.25.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
websocket-client>=1.7.0
